+++ tests/platform/broadcom/test_ser.py
@@ -20,49 +20,10 @@ FILES_DIR = os.path.join(BASE_DIR, 'files')
 SER_INJECTOR_FILE = 'ser_injector.py'
 DUT_WORKING_DIR = '/tmp/'
 
-def disable_ssh_timout(dut):
-    '''
-    @summary disable ssh session on target dut
-    @param dut: Ansible host DUT
-    '''
-    logger.info('Disabling ssh time out on dut: %s' % dut.hostname)
-    dut.command("sudo sed -i 's/^ClientAliveInterval/#&/' /etc/ssh/sshd_config")
-    dut.command("sudo sed -i 's/^ClientAliveCountMax/#&/' /etc/ssh/sshd_config")
-
-    dut.command("sudo systemctl restart ssh")
-    time.sleep(5)
-
-def enable_ssh_timout(dut):
-    '''
-    @summary: enable ssh session on target dut
-    @param dut: Ansible host DUT
-    '''
-    logger.info('Enabling ssh time out on dut: %s' % dut.hostname)
-    dut.command("sudo sed -i '/^#ClientAliveInterval/s/^#//' /etc/ssh/sshd_config")
-    dut.command("sudo sed -i '/^#ClientAliveCountMax/s/^#//' /etc/ssh/sshd_config")
-
-    dut.command("sudo systemctl restart ssh")
-    time.sleep(5)
-
-@pytest.fixture(scope='module')
-def setup(testbed_devices):
-    '''
-    @summary: Test fixture for SER injection test. SER injection test is time consuming.
-              This result in ssh session timing out. The setup test fixture disable ssh 
-              time out during the test and restores it after test is complete
-    @param testbed_devices: Ansible framework testbed devices
-    '''
-    dut = testbed_devices["dut"]
-
-    disable_ssh_timout(dut)
-
-    yield
-
-    enable_ssh_timout(dut)
-
 @pytest.mark.disable_loganalyzer
 @pytest.mark.broadcom
-def test_ser(testbed_devices, setup):
+@pytest.mark.usefixtures("pause_testbed_ssh_timeout")
+def test_ser(testbed_devices, pause_testbed_ssh_timeout):
     '''
     @summary: Broadcom SER injection test use Broadcom SER injection utility to insert SER
               into different memory tables. Before the SER injection, Broadcom mem/sram scanners
